requirements.txt
flask
flask-sqlalchemy
pillow
requests
celery
redis